mslogic: Fix and simplify block pushing logic.
Try to push blocks /first/ before checking what's under them.  This
allows us to remove the recusive call to canmakemove and the associated
CMM_NOPUSHING flag.

Fixes some behaviour when a block is in the lower layer. For example, if
Chip enters a tile containing a key on top of a block, MSCC puts Chip on
top of the block, but in tworld Chip would push the block. (In both
cases Chip picks up the key.)